fix: add PluginOutput#text(); alias #string(); incl in docs (#37)
I flubbed naming the string output function: it should be `.text()`
in order to align with the Web Platform `Response` type, but for
some reason, I called it `.string()` instead. This corrects that error,
while preserving the `.string()` method as a hidden item so it's not
a breaking change.

Additionally, export the `PluginOutput` type so it's visible in typedoc.