feat: implement worker timeout
This applies the timeout to both the initialization of the plugin and execution
of plugin methods. We have to be a little careful about leaving `MessagePort`
instances alive in Node.

It looks like the promise returned by `worker.terminate()` doesn't
ref() properly in Bun. We can work around this by keeping a timer
alive while we terminate the worker process.

However, that reveals that Bun appears not to be able to handle
halting Wasm execution in a thread. So -- for now, at least --
disable timeouts in Bun.

Additionally: accept manifest options in snakeCase and accept memory settings
via plugin options.

Fixes #74.