!squash bun workers don't deal with termination well
It looks like the promise returned by `worker.terminate()` doesn't
ref() properly in Bun. We can work around this by keeping a timer
alive while we terminate the worker process.

However, that reveals that Bun appears not to be able to handle
halting Wasm execution in a thread. So -- for now, at least --
disable timeouts in Bun.